requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.2
billiard==4.1.0
black==23.3.0
blinker==1.6.2
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
cron-descriptor==1.4.0
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==4.0.1
Django==4.2.1
django-admin-rangefilter==0.13.0
django-allauth==0.54.0
django-celery-beat==2.5.0
django-cors-headers==4.1.0
django-countries==7.5.1
django-extensions==3.2.3
django-filter==23.2
django-import-export
django-model-utils==4.3.1
django-phonenumber-field==7.1.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
ecs-logging==2.1.0
elastic-apm
Faker==12.0.1
fire==0.5.0
firebase-admin==6.5.0
flake8==6.0.0
Flask==2.3.2
Flask-Cors==4.0.1
Flask-Login==0.6.3
flask-restx==1.3.0
fuzzywuzzy
gevent==24.2.1
geventhttpclient==2.0.12
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.4
importlib-resources==5.12.0
importlib_metadata==7.1.0
inflection==0.5.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.3.1
locust==2.25.0
MarkupSafe==2.1.3
mccabe==0.7.0
mixer==7.2.2
msgpack==1.0.8
mypy-extensions==1.0.0
Naked==0.1.32
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
paytm==0.1.8
paytmchecksum==1.7.0
phonenumbers==8.13.11
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.10.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.20.0
pycryptodomex==3.17
pyflakes==3.0.1
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
pyzmq==26.0.3
redis~=5.0.7
geopy==2.4.1
requests==2.31.0
requests-oauthlib==1.3.1
responses~=0.25.3
rsa==4.9
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tablib==3.5.0
termcolor==2.3.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.16.0
zipp==3.15.0
zope.event==5.0
zope.interface==6.4.post2
